requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-rest-auth==1.1.0
Django==4.1.2
djangorestframework==3.14.0
easy-thumbnails==2.8.3
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.4
sqlparse==0.4.3
